// src/types/partials/PaymentReconciliationDetail.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDate } from '../simpleTypes/Date';
import { TMoney } from '../partials/Money';

export type TPaymentReconciliationDetail = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    identifier?: TIdentifier;
    predecessor?: TIdentifier;
    type: TCodeableConcept;
    request?: TReference;
    submitter?: TReference;
    response?: TReference;
    date?: TDate;
    responsible?: TReference;
    payee?: TReference;
    amount?: TMoney;
};